chore: move state storage code to store.go, rm ambiguous wasm.go and wasm_test.go files
