+++ modules/light-clients/08-wasm/types/store.go
@@ -2,11 +2,13 @@ package types
 
 import (
 	"bytes"
+	"context"
 	"errors"
 	"io"
 	"reflect"
 	"strings"
 
+	wasmvm "github.com/CosmWasm/wasmvm"
 	wasmvmtypes "github.com/CosmWasm/wasmvm/types"
 
 	errorsmod "cosmossdk.io/errors"
@@ -14,6 +16,8 @@ import (
 	storeprefix "cosmossdk.io/store/prefix"
 	"cosmossdk.io/store/tracekv"
 	storetypes "cosmossdk.io/store/types"
+
+	"github.com/cosmos/ibc-go/modules/light-clients/08-wasm/internal/ibcwasm"
 )
 
 var (
@@ -23,6 +27,50 @@ var (
 	subjectPrefix    = []byte("subject/")
 	substitutePrefix = []byte("substitute/")
 )
+
+// Checksum is a type alias used for wasm byte code checksums.
+type Checksum = wasmvmtypes.Checksum
+
+// CreateChecksum creates a sha256 checksum from the given wasm code, it forwards the
+// call to the wasmvm package. The code is checked for the following conditions:
+// - code length is zero.
+// - code length is less than 4 bytes (magic number length).
+// - code does not start with the wasm magic number.
+func CreateChecksum(code []byte) (Checksum, error) {
+	return wasmvm.CreateChecksum(code)
+}
+
+// GetAllChecksums is a helper to get all checksums from the store.
+// It returns an empty slice if no checksums are found
+func GetAllChecksums(ctx context.Context) ([]Checksum, error) {
+	iterator, err := ibcwasm.Checksums.Iterate(ctx, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	keys, err := iterator.Keys()
+	if err != nil {
+		return nil, err
+	}
+
+	checksums := []Checksum{}
+	for _, key := range keys {
+		checksums = append(checksums, key)
+	}
+
+	return checksums, nil
+}
+
+// HasChecksum returns true if the given checksum exists in the store and
+// false otherwise.
+func HasChecksum(ctx context.Context, checksum Checksum) bool {
+	found, err := ibcwasm.Checksums.Has(ctx, checksum)
+	if err != nil {
+		return false
+	}
+
+	return found
+}
 
 // migrateClientWrappedStore combines two KVStores into one.
 //
@@ -205,22 +253,27 @@ func newStoreAdapter(s storetypes.KVStore) *storeAdapter {
 	return &storeAdapter{parent: s}
 }
 
+// Get implements the KVStore interface.
 func (s storeAdapter) Get(key []byte) []byte {
 	return s.parent.Get(key)
 }
 
+// Set implements the KVStore interface.
 func (s storeAdapter) Set(key, value []byte) {
 	s.parent.Set(key, value)
 }
 
+// Delete implements the KVStore interface.
 func (s storeAdapter) Delete(key []byte) {
 	s.parent.Delete(key)
 }
 
+// Iterator implements the KVStore interface.
 func (s storeAdapter) Iterator(start, end []byte) wasmvmtypes.Iterator {
 	return s.parent.Iterator(start, end)
 }
 
+// ReverseIterator implements the KVStore interface.
 func (s storeAdapter) ReverseIterator(start, end []byte) wasmvmtypes.Iterator {
 	return s.parent.ReverseIterator(start, end)
 }
